requirements.txt
faiss-cpu
numpy
pandas
PyYAML
requests
scikit-learn 